pyproject.toml
[tool.poetry]
name = "dbt-eppo-sync"
version = "0.6.0"
description = "A package to sync dbt semantic layer definitions to Eppo."
authors = ["Anthony Rindone <anthony@geteppo.com>"] 
readme = "README.md"
license = "MIT"
packages = [{include = "dbt_eppo_sync", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9" 
pyyaml = "^6.0" # For parsing YAML files
requests = "^2.31.0" # For making API calls to Eppo
dbt-core = "^1.8.0" # For interacting with dbt artifacts
click = "^8.1.8"
jsonschema = "^4.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4" # For running tests

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dbt-eppo-sync = "dbt_eppo_sync.cli:main"

